apple-silicon: import native arm64 port onto clean main
Applies the port from native-smoke-v1 (commit 4ee449fda) onto upstream
db8fd17e3, excluding the C0 evidence-contract docs and scripts.

Engine + build port is ~10k lines (src/storage 5.5k, src/unit_test 1.6k,
src/CMakeLists 508, src/common/simd 404, cmake/vcpkg/ports); the remaining
~93k lines are the tools/apple_silicon and tools/portability benchmark and
smoke harnesses, which are pruned in follow-up commits.

Imported as-is as a known-good starting point. Not yet simplified.

+++ src/storage/knn_index/plaid/plaid_quantizer_impl.cpp
@@ -152,7 +152,7 @@ void PlaidQuantizer::Train(const f32 *residuals, u64 n_embeddings) {
     bucket_cutoffs_ = std::make_unique<f32[]>(n_buckets_ - 1);
     for (u32 i = 1; i < n_buckets_; ++i) {
         f64 quantile = static_cast<f64>(i) / n_buckets_;
-        u64 idx = static_cast<u64>(quantile * all_residuals.size());
+        size_t idx = static_cast<size_t>(quantile * all_residuals.size());
         idx = std::min(idx, all_residuals.size() - 1);
         bucket_cutoffs_[i - 1] = all_residuals[idx];
     }
@@ -163,7 +163,7 @@ void PlaidQuantizer::Train(const f32 *residuals, u64 n_embeddings) {
     bucket_weights_ = std::make_unique<f32[]>(n_buckets_);
     for (u32 i = 0; i < n_buckets_; ++i) {
         f64 quantile = (static_cast<f64>(i) + 0.5) / n_buckets_;
-        u64 idx = static_cast<u64>(quantile * all_residuals.size());
+        size_t idx = static_cast<size_t>(quantile * all_residuals.size());
         idx = std::min(idx, all_residuals.size() - 1);
         bucket_weights_[i] = all_residuals[idx] + avg_residual_;
     }
